refactor: Enforce shared-engine parity for check_phase_handoff (#775)
* refactor: move phase handoff simulation and analysis logic to shared engine

Delegates the inline Phase 1 equilibration simulation, conservation law analysis, Phase 2 discontinuity checking, and dynamic response evaluation from handleCheckPhaseHandoff to the shared-engine service analyzePhaseHandoff in packages/engine/src/services/analysis/PhaseHandoff.ts to enforce strict shared-engine parity.


* chore: auto-fix linting and formatting issues

* refactor: move phase handoff simulation and analysis logic to shared engine

Delegates the inline Phase 1 equilibration simulation, conservation law analysis, Phase 2 discontinuity checking, and dynamic response evaluation from handleCheckPhaseHandoff to the shared-engine service analyzePhaseHandoff in packages/engine/src/services/analysis/PhaseHandoff.ts to enforce strict shared-engine parity.


* chore: auto-fix linting and formatting issues

---------
